+++ CHANGES.rst
@@ -5,7 +5,10 @@ Changelog of lizard-auth-server
 2.27 (unreleased)
 -----------------
 
-- Nothing changed yet.
+- Transparently use cognito instead of our own DB/AD for authentication. This
+  way, sites can gradually switch to using cognito. Some use it directly,
+  others use it transparently through the SSO. This is a temporary situation
+  before the SSO is phased out.
 
 
 2.26 (2020-11-27)